scale_value::stringify

Module custom_formatters

This module provides custom formatters that work alongside crate::stringify::to_writer_custom and allow for the output to be formatted in various ways.

Functions

format_hex

This can be used alongside crate::stringify::ToWriterBuilder::add_custom_formatter() (which itself is constructed via crate::stringify::to_writer_custom()). It will format as a hex string any unnamed composite whose values are all primitives in the range 0-255 inclusive.